--- packages/apps/mercadopago/firebase.ts
@@ -1,2 +1,0 @@
-/* eslint-disable */ 
-export * from './lib/firebase'
