+++ packages/apps/mercadopago/scripts/build.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-rm firebase.ts
-node ../../../scripts/public-apps-minification.mjs mercadopago 
+node ../../../scripts/script-minification.mjs
 sh ../../../scripts/build-lib.sh
 
-echo "/* eslint-disable */ 
-export * from './lib/firebase'" >> firebase.ts
