+++ packages/apps/mercadopago/src/mercadopago.ts
@@ -4,8 +4,8 @@ import type { AppModuleBody } from '@cloudcommerce/types';
 // eslint-disable-next-line import/no-unresolved
 import { getFirestore } from 'firebase-admin/firestore';
 // eslint-disable-next-line import/no-unresolved
-import handleListPayments from './modules/list-payments';
-import handleCreateTransaction from './modules/create-transaction';
+import handleListPayments from './mp-list-payments';
+import handleCreateTransaction from './mp-create-transaction';
 
 export const listPayments = async (modBody: AppModuleBody) => {
   return handleListPayments(modBody);
