+++ packages/apps/mercadopago/webhook.js
@@ -1,0 +1,1 @@
+export * from './lib/mp-webhook.js';
