+++ packages/events/src/firebase.ts
@@ -3,4 +3,4 @@ import '@cloudcommerce/firebase/lib/init';
 
 export * from '@cloudcommerce/app-tiny-erp';
 
-export * from '@cloudcommerce/app-mercadopago/firebase';
+export * from '@cloudcommerce/app-mercadopago/webhook';
